requirements.txt
numpy==1.21.5
Pillow==9.0.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
pathlib==1.0.1
packaging
random
